Build Room object with messages and users in it vs hitting API all the time
