--- user.py
@@ -1,7 +1,0 @@
-__author__ = 'Chad Peterson'
-__email__ = 'chapeter@cisco.com'
-
-class User(object):
-    """
-    Class used to create Users
-    """
